perf: streamlined design
