+++ src/shared/presenter.d.ts
@@ -1814,3 +1814,116 @@ export interface IVectorDatabasePresenter {
    */
   resumeAllPausedTasks(): Promise<void>
 }
+
+/**
+ * Lifecycle management types and interfaces
+ */
+
+import { ISQLitePresenter } from '@shared/presenter'
+
+/**
+ * Context object passed to lifecycle hooks during execution
+ */
+export interface LifecycleContext {
+  phase: LifecyclePhase
+  manager: ILifecycleManager
+  database?: ISQLitePresenter
+  config?: any
+  [key: string]: any
+}
+
+/**
+ * Lifecycle hook interface for components to register phase-specific logic
+ */
+export interface LifecycleHook {
+  name: string // Descriptive name for logging and debugging
+  priority?: number // Lower numbers execute first (default: 100)
+  execute: (context: LifecycleContext) => Promise<void | boolean>
+  timeout?: number // Optional timeout in milliseconds
+  critical?: boolean // If true, failure halts the phase (default: false)
+}
+
+/**
+ * Internal lifecycle state tracking
+ */
+export interface LifecycleState {
+  currentPhase: LifecyclePhase | null
+  completedPhases: Set<LifecyclePhase>
+  startTime: number
+  phaseStartTimes: Map<LifecyclePhase, number>
+  hooks: Map<LifecyclePhase, Array<{ id: string; hook: LifecycleHook }>>
+  isShuttingDown: boolean
+}
+
+/**
+ * LifecycleManager interface defining the core lifecycle management API
+ */
+export interface ILifecycleManager {
+  // Phase management
+  start(): Promise<void>
+  getCurrentPhase(): LifecyclePhase | null
+  isPhaseComplete(phase: LifecyclePhase): boolean
+
+  // Hook registration - for components that need to execute logic during specific phases
+  registerHook(phase: LifecyclePhase, hook: LifecycleHook): string // Returns generated hook ID
+  unregisterHook(phase: LifecyclePhase, hookId: string): void
+
+  // Shutdown control
+  requestShutdown(): Promise<boolean>
+  forceShutdown(): void
+
+  // Splash window management
+  getSplashManager(): ISplashWindowManager
+
+  // Error handling and recovery
+  getErrorStatistics(): {
+    totalErrors: number
+    criticalErrors: number
+    nonCriticalErrors: number
+    retriedHooks: number
+    failedHooks: Array<{
+      hookName: string
+      phase: LifecyclePhase
+      error: string
+      timestamp: number
+    }>
+  }
+  updateRetryConfig(config: {
+    maxRetries?: number
+    retryDelay?: number
+    backoffMultiplier?: number
+    maxRetryDelay?: number
+  }): void
+  clearErrorHistory(): void
+
+  // Context management
+  getLifecycleContext(): LifecycleContext
+}
+
+export interface ISplashWindowManager {
+  create(): Promise<void>
+  updateProgress(phase: LifecyclePhase, progress: number): void
+  updateMessage(message: string): void
+  close(): Promise<void>
+  isVisible(): boolean
+}
+
+export interface LifecycleEventStats {
+  totalPhases: number
+  completedPhases: number
+  totalHooks: number
+  successfulHooks: number
+  failedHooks: number
+  totalDuration: number
+  phaseStats: Map<
+    string,
+    {
+      duration: number
+      hookCount: number
+      successfulHooks: number
+      failedHooks: number
+      startTime: number
+      endTime: number
+    }
+  >
+}
